chore(js/genkit): make beta export a superset of stable instead of a diff
