+++ js/genkit/src/index.ts
@@ -1,147 +1,2 @@
-/**
- * @license
- *
- * Copyright 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Main genkit import.
- *
- * ```ts
- * import { genkit } from 'genkit';
- * ```
- *
- * @module /
- */
-export {
-  BaseDataPointSchema,
-  Document,
-  DocumentDataSchema,
-  GenerationBlockedError,
-  GenerationCommonConfigSchema,
-  GenerationResponseError,
-  LlmResponseSchema,
-  LlmStatsSchema,
-  Message,
-  MessageSchema,
-  ModelRequestSchema,
-  ModelResponseSchema,
-  PartSchema,
-  RoleSchema,
-  ToolCallSchema,
-  ToolSchema,
-  embedderRef,
-  evaluatorRef,
-  indexerRef,
-  rerankerRef,
-  retrieverRef,
-  type DocumentData,
-  type EmbedderAction,
-  type EmbedderArgument,
-  type EmbedderInfo,
-  type EmbedderParams,
-  type EmbedderReference,
-  type Embedding,
-  type EvalResponses,
-  type EvaluatorAction,
-  type EvaluatorInfo,
-  type EvaluatorParams,
-  type EvaluatorReference,
-  type GenerateOptions,
-  type GenerateRequest,
-  type GenerateRequestData,
-  type GenerateResponse,
-  type GenerateResponseChunk,
-  type GenerateResponseChunkData,
-  type GenerateResponseData,
-  type GenerateStreamOptions,
-  type GenerateStreamResponse,
-  type GenerationUsage,
-  type IndexerAction,
-  type IndexerArgument,
-  type IndexerInfo,
-  type IndexerParams,
-  type IndexerReference,
-  type InterruptConfig,
-  type LlmResponse,
-  type LlmStats,
-  type MediaPart,
-  type MessageData,
-  type ModelArgument,
-  type ModelReference,
-  type ModelRequest,
-  type ModelResponseData,
-  type Part,
-  type PromptAction,
-  type PromptConfig,
-  type RankedDocument,
-  type RerankerAction,
-  type RerankerArgument,
-  type RerankerInfo,
-  type RerankerParams,
-  type RerankerReference,
-  type ResumeOptions,
-  type RetrieverAction,
-  type RetrieverArgument,
-  type RetrieverInfo,
-  type RetrieverParams,
-  type RetrieverReference,
-  type Role,
-  type Tool,
-  type ToolAction,
-  type ToolArgument,
-  type ToolCall,
-  type ToolConfig,
-  type ToolRequestPart,
-  type ToolResponsePart,
-} from '@genkit-ai/ai';
-export { Chat } from '@genkit-ai/ai/chat';
-export {
-  Session,
-  type SessionData,
-  type SessionStore,
-} from '@genkit-ai/ai/session';
-export {
-  GENKIT_CLIENT_HEADER,
-  GENKIT_VERSION,
-  GenkitError,
-  ReflectionServer,
-  StatusCodes,
-  StatusSchema,
-  defineJsonSchema,
-  defineSchema,
-  getCurrentEnv,
-  getStreamingCallback,
-  isDevEnv,
-  runWithStreamingCallback,
-  z,
-  type Action,
-  type ActionContext,
-  type ActionMetadata,
-  type Flow,
-  type FlowConfig,
-  type FlowFn,
-  type FlowSideChannel,
-  type JSONSchema,
-  type JSONSchema7,
-  type Middleware,
-  type ReflectionServerOptions,
-  type RunActionResponse,
-  type Status,
-  type StreamingCallback,
-  type StreamingResponse,
-  type TelemetryConfig,
-} from '@genkit-ai/core';
+export * from './common.js';
 export { Genkit, genkit, type GenkitOptions } from './genkit.js';
